Remove eventInterval
Events in #314 can now be defined in numerous ways, not only as the maximum duration between timestamps. We could make the field not required, but after discussion with @kbubnicki we think it's better to remove it altogether to not give particular weight to this method. It can also to a large extend be derived from the data.