Shorten UUIDs to first part
They are still unique, but much easier to read in examples. It will break their link to Agouti, but the example has moved on quite for from the Agouti export. Very likely it will still be possible to find the data in Agouti based on the first part of the UUID.